+++ packages/@runlightyear/slack/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @runlightyear/slack
 
+## 0.11.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @runlightyear/lightyear@0.18.0
+
+## 0.11.0
+
+### Minor Changes
+
+- Minor updates to Slack elements and documentation
+
 ## 0.10.0
 
 ### Minor Changes